move components to web package (#37)
* move components

* fix
