+++ packages/web/components/common/DateRangePicker/index.tsx
@@ -2,10 +2,10 @@ import React, { useState, useMemo, useRef } from 'react';
 import { Box, Card, Flex, useTheme, useOutsideClick, Button } from '@chakra-ui/react';
 import { addDays, format } from 'date-fns';
 import { type DateRange, DayPicker } from 'react-day-picker';
-import MyIcon from '@fastgpt/web/components/common/Icon';
 import 'react-day-picker/dist/style.css';
 import zhCN from 'date-fns/locale/zh-CN';
 import { useTranslation } from 'next-i18next';
+import MyIcon from '../Icon';
 
 const DateRangePicker = ({
   onChange,
