+++ packages/web/package.json
@@ -26,9 +26,15 @@
     "@lexical/react": "0.12.6",
     "papaparse": "^5.4.1",
     "@lexical/utils": "0.12.6",
-    "@lexical/text": "0.12.6"
+    "@lexical/text": "0.12.6",
+    "date-fns": "2.30.0",
+    "react-day-picker": "^8.7.1",
+    "lodash": "^4.17.21",
+    "@tanstack/react-query": "^4.24.10",
+    "dayjs": "^1.11.7"
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.191",
     "@types/react": "18.2.0",
     "@types/papaparse": "^5.3.7",
     "@types/react-dom": "18.2.0",
