+++ projects/app/src/components/Select/SelectAiModel.tsx
@@ -1,10 +1,10 @@
 import React, { useMemo } from 'react';
 
-import MySelect, { type SelectProps } from './index';
 import { useTranslation } from 'next-i18next';
 import { useSystemStore } from '@/web/common/system/useSystemStore';
 import { useRouter } from 'next/router';
 import { AI_POINT_USAGE_CARD_ROUTE } from '@/web/support/wallet/sub/constants';
+import MySelect, { SelectProps } from '@fastgpt/web/components/common/MySelect';
 
 const SelectAiModel = ({ list, ...props }: SelectProps) => {
   const { t } = useTranslation();
