+++ projects/app/src/components/common/MyBox/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Box, BoxProps } from '@chakra-ui/react';
-import Loading from '@/components/Loading';
+import Loading from '@fastgpt/web/components/common/MyLoading';
 
 type Props = BoxProps & {
   isLoading?: boolean;
